@netrunnereve 2024-06-16
> Make sure the `test-backend-ops` passes

Yeah it runs and passes with `-b CPU`; interestingly enough the test is skipped by default if no backend is set. I actually used `test-quantize-fns` when writing this PR.
```
./tests/test-backend-ops -b CPU
Testing 1 backends

Backend 1/1 (CPU)
  Backend name: CPU
  ABS(type=f32,ne_a=[128,10,10,10],v=0): OK
...
  FLASH_ATTN_EXT(hs=256,nh=32,kv=1024,nb=8,mask=0,max_bias=0.000000,type_KV=q4_0): OK
  1270/1270 tests passed
  Backend CPU: OK

1/1 backends passed
OK
```
Considering how it only takes a minute to run I think it's worth adding the CPU version of `test-backend-ops` to the CI.