@slaren 2024-06-16
The goal of `test-backend-ops` is to compare GPU backends with the CPU backend. When used with the CPU backend, it only compares the CPU backend with itself. It can still detect some issues such as `nan` and `inf` values, but generally I don't think it adds much value.