@ggerganov 2024-06-18
I compared the AVX CPU vs the GPU results on my linux box and tests are passing. Should be good to merge